faas/text-core-phrase-extractor/requirements.txt
sumy
nltk
numpy
guess_language-spirit